Refactor interpreter a bit
* Create src/interp/ directory
* Move Thread::Trace into src/interp/interp-trace.cc
* Move Environment::Disassemble into src/interp/interp-disassemble.cc
* Move shared code into src/interp/interp-internal.h
