+++ src/test-interp.cc
@@ -23,9 +23,9 @@
 #include <vector>
 
 #include "src/binary-reader.h"
-#include "src/binary-reader-interp.h"
 #include "src/cast.h"
-#include "src/interp.h"
+#include "src/interp/binary-reader-interp.h"
+#include "src/interp/interp.h"
 #include "src/make-unique.h"
 
 using namespace wabt;
